apps/api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
openai>=1.51
python-multipart==0.0.6
Pillow==10.1.0
distro==1.9.0
langgraph>=0.2.40
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9
sqlalchemy>=2.0
alembic>=1.13.0
bcrypt==4.0.1
email-validator>=2.0.0
APScheduler==3.10.4
python-dateutil==2.8.2
pytz>=2023.3

authlib

pgvector==0.2.4

itsdangerous

jwt
pymongo[srv]>=4.7
